tests/test_yolo_integration.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
YOLO集成测试模块
测试YOLOv8/v10模型集成和预测功能
"""

import unittest
import tempfile
import os
import sys
import numpy as np
from PIL import Image

sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..'))

try:
    from visiofirm.utils.VFPreAnnotator import ImageProcessor
    IMPORTS_OK = True
except ImportError as e:
    print(f"警告: 无法导入VFPreAnnotator: {e}")
    IMPORTS_OK = False


class TestYOLOIntegration(unittest.TestCase):
    """YOLO集成测试类"""
    
    @classmethod
    def setUpClass(cls):
        """测试类设置"""
        if not IMPORTS_OK:
            cls.skipTest(cls, "VFPreAnnotator模块导入失败")
        
        cls.temp_dir = tempfile.mkdtemp()
        
        # 创建测试图像
        cls.test_image_path = os.path.join(cls.temp_dir, 'test_image.jpg')
        test_img = Image.new('RGB', (640, 480), color='red')
        test_img.save(cls.test_image_path)
    
    @classmethod
    def tearDownClass(cls):
        """测试类清理"""
        if os.path.exists(cls.temp_dir):
            import shutil
            shutil.rmtree(cls.temp_dir)
    
    def test_yolo_model_initialization_cpu(self):
        """测试YOLO模型CPU初始化"""
        if not IMPORTS_OK:
            self.skipTest("VFPreAnnotator模块不可用")
        
        try:
            processor = ImageProcessor(
                model_type="yolo",
                yolo_model_path="yolov8n.pt",
                device="cpu"
            )
            self.assertIsNotNone(processor)
        except Exception as e:
            self.skipTest(f"YOLO模型初始化失败: {e}")
    
    def test_yolo_predict_single_image(self):
        """测试YOLO单张图像预测"""
        if not IMPORTS_OK:
            self.skipTest("VFPreAnnotator模块不可用")
        
        self.skipTest("需要实际图像和模型权重")
    
    def test_yolo_result_parsing(self):
        """测试YOLO结果解析"""
        # 测试结果格式是否正确
        pass
    
    def test_yolo_bbox_conversion(self):
        """测试YOLO边界框坐标转换"""
        # 测试YOLO格式到标准格式的转换
        pass
    
    def test_yolo_confidence_filtering(self):
        """测试置信度过滤"""
        # 测试低置信度结果被过滤
        pass


if __name__ == '__main__':
    unittest.main(verbosity=2)
