requirements.txt
filelock==3.19.1
Flask==3.1.2
Flask_Login==0.6.3
networkx==3.4.2
openai_clip==1.0.1
opencv_python==4.12.0.88
Pillow==11.3.0
psutil==7.0.0
PyYAML==6.0.2
rapidfuzz==3.13.0
rarfile==4.2
requests>=2.31.0
torch==2.8.0
tqdm==4.66.1
transformers==4.55.4
ultralytics==8.3.185
Werkzeug==3.1.3
groundingdino-py
uvicorn==0.32.0
waitress==3.0.2
